project_data/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
requests
psutil
python-multipart
redis
aioredis
